refactor(pages): introduce `FollowedContent` component for displaying followed production company content on movie and TV pages with filtering and sorting.
